% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prob_convenience.R
\name{coinflipper}
\alias{coinflipper}
\title{coinflipper}
\usage{
coinflipper(n.flips = 1, n.success = 1, p = 0.5)
}
\arguments{
\item{n.flips}{}

\item{n.success}{}

\item{p}{}
}
\description{
simulates coin flips
}

