% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{ct}
\alias{ct}
\title{ct}
\usage{
ct(df = 10, type = "equal", conf = 0.95)
}
\arguments{
\item{df}{(int) degrees of freedom.}

\item{type}{string. type of hypothesis test taken.

       - for two tailed test use "equal" (Default)  or "="

       - for one-tailed lower tail test use "less" or "<"

       - for one-tailed upper tail test use "more" or "greater" or "higher"
         or ">"}

\item{conf}{confidence interval used}
}
\value{
a vector with two values representing the lower and upper points that
        fit within your confidence interval.
}
\description{
Confidence region for a T Distribution.
}
\examples{
ct(df=20, type="equal", conf=0.99)
ct(df=20, type="=", conf=0.90)
ct(df=20, type="less", conf=0.95)
ct(df=20, type="<", conf=0.90)
ct(df=20, type="more", conf=0.90)
ct(df=20, type=">", conf=0.90)
}

