% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error.R
\name{error}
\alias{error}
\title{error}
\usage{
error(x.pred, x.actual, type = c("MSE", "RMSE"))
}
\arguments{
\item{x.pred}{(numeric) predicted value(s)}

\item{x.actual}{(numeric) actual values}

\item{type}{(string) type of error function to use.

         "MSE"  Mean Squared Error

         "RMSE" Root Mean Squared Error}
}
\description{
calculates the error value for some vector of values
}
\details{
calculates the error value for some vector of values
}
\examples{
mu = mean(mtcars$mpg)
error(mu, mtcars$mpg, type="MSE")
}
\author{
Ronny Restrepo
}
\keyword{error,}
\keyword{residuals,}
\keyword{stat.convenience}
\keyword{stats,}

