% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataframe_convenience.R
\name{filter.columns}
\alias{filter.columns}
\title{filter.columns}
\usage{
filter.columns(x, pattern, exclude = FALSE, ...)
}
\arguments{
\item{x}{(dataframe) The dataframe you want to filter.}

\item{pattern}{(string) A regular expression.}

\item{exclude}{(logical) Should the columns matching the pattern be filtered
       out? If TRUE, then it returns all columns except the ones matching the
       pattern.. If FALSE, then it returns only the columns that match.

        (DEFAULT = FALSE)}

\item{...}{aditional parameters to be passed on to grep()}
}
\description{
Returns a copy of a dataframe which filters for, or filters out, only the
columns that fit some regular expression pattern. You can chose if the
columns that match the pattern should be the ones returned, or if they are
the ones that should be excluded.
}
\examples{
# Returns only the columns that start with "Petal"
data(iris)
filter.columns(iris, "^Petal")

# Filters out the columns that start with "Petal"
filter.columns(iris, "^Petal", exclude=TRUE)
}
\author{
Ronny Restrepo
}

