% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat_convenience.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(x)
}
\arguments{
\item{x}{(vector of numerics) The values}
}
\value{
(numeric) the weighted mean
}
\description{
Takes a vector of numerics, and normalizes the data so that we end up with a
mean of 0 and standard deviation of 1.
}
\examples{
x = c(12,14,11,16)
x.norm = normalize(x)
x.norm        # -0.5637345  0.3382407 -1.0147221  1.2402159
mean(x.norm)  # 0
sd(x.norm)    # 1
}
\author{
Ronny Restrepo
}

