% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat_convenience.R
\name{wmean}
\alias{wmean}
\title{wmean}
\usage{
wmean(x, w = NA)
}
\arguments{
\item{x}{(vector of numerics) The values}

\item{w}{(vector of numerics) The corresponding weights

         If NA, then it defaults to weights of 1.

         (DEFAULT = NA = weights of 1)}
}
\value{
(numeric) the weighted mean
}
\description{
Calculatest the weighted mean from a vector of values (x) and a vector of
corresponding weights (w) for each of the elements in x.
}
\examples{
x = c(3.5, 5.2, 2.7, 4.2)
weights = c(1, 2, 1, 6)
wmean(x, weights)        # 4.18
mean(x)                  # 3.9
}
\author{
Ronny Restrepo
}

