% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{cbinom}
\alias{cbinom}
\title{cbinom}
\usage{
cbinom(size = 1, p = 0.5, type = "equal", conf = 0.95)
}
\arguments{
\item{size}{int. number of trials for the binomial distribution.}

\item{p}{numeric. probability of success per trial}

\item{type}{string. type of hypothesis test taken.
"equal" (Default) for two tailed test
"less" for one-tailed test where alternative hypotheis is 'less than'
"more" for one-tailed test where alternative hypotheis is 'more than'}

\item{conf}{confidence interval used}
}
\value{
a vector with two values representing the lower and upper points that
        fit within your confidence interval.
}
\description{
Confidence region for a binomial Distribution.
}
\examples{
cbinom(size=100, p=0.7, type="equal", conf=0.99)
cbinom(size=15, p=0.9, type="less", conf=0.95)
cbinom(size=30, p=0.4, type="more", conf=0.90)
}

