% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat_convenience.R
\name{lm2}
\alias{lm2}
\title{lm2}
\usage{
lm2(x, y, print.summary = TRUE)
}
\arguments{
\item{x}{(vector of numerics) The independent variables}

\item{y}{(vector of numerics) The dependent variables}

\item{print.summary}{(boolean) Should it print a summary?
(DEFAULT = TRUE)}
}
\value{
a list of values:

     $n

     $mean.x

     $mean.y

     $sd.x

     $sd.y

     $cor

     $rot_significance

     $rot_is_significant

     $slope

     $intercept

     $SST                # Total Sum of Squared Errors (using mean of y)

     $SSE                # Sum of Squared Errors in the Model

     $SSR                # Sum of Squared Errors due to Regression

     $MST                # Total Mean Squared Errors (using mean of y)

     $MSE                # Mean Squared Errors in the model

     $cod                # Coefficient of Determination
}
\description{
Creates a list of information needed for creating a linear model, along with
other information that is useful in evaluating how good that model is.
}
\examples{
lm2(x, y)
lm2(x, y, print.summary=FALSE)

# Plot a scatterplot and the linear model
model = lm2(x,y)
plot(x, y)
abline(model$intercept, model$slope)
}
\author{
Ronny Restrepo
}

