% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataframe_convenience.R
\name{na.summary}
\alias{na.summary}
\title{na.summary}
\usage{
na.summary(df, lineup = TRUE, limit = 8, thresh = NA, only.nas = FALSE,
  no.nas = FALSE, printit = TRUE)
}
\arguments{
\item{df}{(dataframe) The dataframe you want to check for NAs}

\item{lineup}{(boolean) Try to line things up nicely. By default it
is set to TRUE, but if you have one column name that is much longer
than all the others it may cause the values to be pushed too far to
the right and may make it look ugly.
FALSE = disables this automatic lining up feature.
(DEFAULT = TRUE)}

\item{limit}{(integer) max number of NA indices to print out

        (DEFAULT = 8)}

\item{thresh}{(numerical) threshold value between 0 and 1, representing
proportion of NAs in the column. So iff you use thresh=0.3 it
returns NA information for all the columns that are composed of more
than 30% NAs}

\item{only.nas}{(logical) Only return summaries for columns containing NAs?

         - if FALSE, then it gives summary of all columns.

         - if TRUE, then it only prints out sumary for those columns
           containing at least one NA.

         (DEFAULT=FALSE)}

\item{no.nas}{(logical) prints out only the columns that dont contain any NAs

         - If TRUE, then it only shows the columns that dont contain any NAs.
           This overrides the `only.nas` and `thresh` arguments.

         (DEFAULT = FALSE)}

\item{printit}{(logical) Should it print a pretty version of the summary?

         (DEFAULT = TRUE)}
}
\value{
silently returns a dataframe with the following columns:

         $colName the names of the columns of interest

         $total the total number of NAs in that column

         $proportion the proportion of NAs in that column
}
\description{
Prints a summary of NAs that exist in the dataframe.
}
\examples{
a = c(1,2,3,4,5,6,7,8,9,10)
bravo = c(1,2,NA,NA,5,6,7,8,9,10)
charlie = c(NA,NA,NA,NA,NA,NA,NA,NA,NA,NA)
delta = c(NA,2,3,4,NA,6,7,8,9,NA)
df = data.frame(a,bravo,charlie,delta)
na.summary(df)
#[1] "=========================================="
#[1] "             SUMMARY OF NAS               "
#[1] "=========================================="
#[1] "Column name: Number of NAs: Indices of NAs"
#[1] "                                          "
#[1] "a      : 0: "
#[1] "bravo  : 2: (3, 4)"
#[1] "charlie: 10: (1, 2, 3, 4, 5, 6, 7, 8, ...)"
#[1] "delta  : 3: (1, 5, 10)"
#[1] "=========================================="

na.summary(df, only.nas=TRUE)
#[1] "=========================================="
#[1] "             SUMMARY OF NAS               "
#[1] "=========================================="
#[1] "Column name: Number of NAs: Indices of NAs"
#[1] "                                          "
#[1] "bravo  : 2: (3, 4)"
#[1] "charlie: 10: (1, 2, 3, 4, 5, 6, 7, 8, ...)"
#[1] "delta  : 3: (1, 5, 10)"
#[1] "=========================================="

na.summary(df, no.nas=TRUE)
#[1] "=========================================="
#[1] "             SUMMARY OF NAS               "
#[1] "=========================================="
#[1] "Columns not containing any NAs"
#[1] "                                          "
#[1] "a: 0"
#[1] "=========================================="
}
\author{
Ronny Restrepo
}

