% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataframe_convenience.R
\name{no.nas}
\alias{no.nas}
\title{no.nas}
\usage{
no.nas(df)
}
\arguments{
\item{df}{(dataframe) The dataframe you want to clean up.}
}
\description{
Returns a copy of a dataframe which has all rows containing NAs removed.
}
\examples{
# Create a dataframe
a = c(11,NA,13,14,15,16,17,18,19)
b = c(21,22,NA,24,25,26,27,28,29)
c = c(31,32,33,34,35,36,NA,NA,NA)
d = c(NA,42,43,44,45,46,47,48,49)
df = data.frame(a,b,c,d)

# return only the rows with no NAs in it
no.nas(df)

# Result is
# a  b  c  d
# 4 14 24 34 44
# 5 15 25 35 45
# 6 16 26 36 46
}
\author{
Ronny Restrepo
}

